# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GoFree Surfers API",
        "description": "The GoFree Surfers API is a RESTful API that provides access to data that's collected from the GoFree Surfers database. <br>In Version 1.0.0 this API currently only provides endpoints for maps and players. <br>For more information, please visit the Discord server.<br/><br/>Created by <a href='https://stats.go-free.info/player/STEAM_1:0:50362476'>Simply Brandon</a>",
        "contact": {
            "name": "GoFree Community Discord Server",
            "url": "https://discord.gg/aTf8J7xT9n"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/maps": {
            "get": {
                "tags": [
                    "Maps"
                ],
                "summary": "Get all maps",
                "description": "Returns a paginated response of maps, in the given style.",
                "operationId": "indexMaps",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page of maps to return. Varies giving the amount of maps per page. Default: 1",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "The amount of maps to return per page. Default/Min: 30. Max: 10000.",
                        "schema": {
                            "type": "integer",
                            "maximum": 10000,
                            "minimum": 30
                        }
                    },
                    {
                        "name": "mapname",
                        "in": "query",
                        "description": "Queries the maps names for likeliness to the search term.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tier",
                        "in": "query",
                        "description": "Filters the maps by tier.",
                        "schema": {
                            "type": "integer",
                            "maximum": 8,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Map"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "/api/v1/maps?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 10000
                                        },
                                        "last_page_url": {
                                            "type": "integer",
                                            "example": "/api/v1/maps?page=100"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "/api/v1/maps?page=2"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "/api/v1/maps"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 30
                                        },
                                        "prev_page_url": {
                                            "type": "integer",
                                            "example": null
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Maps not found"
                    },
                    "400": {
                        "description": "Out of Bounds"
                    }
                }
            }
        },
        "/api/v1/maps/{mapname}": {
            "get": {
                "tags": [
                    "Maps"
                ],
                "summary": "Get a map by mapname",
                "description": "Returns a record of a map, if found.",
                "operationId": "showMap",
                "parameters": [
                    {
                        "name": "mapname",
                        "in": "path",
                        "description": "The name of the map to return.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "surf_utopia_v3"
                        }
                    },
                    {
                        "name": "style",
                        "in": "query",
                        "description": "The style of the record to return. Default: 0",
                        "schema": {
                            "type": "integer",
                            "maximum": 10,
                            "minimum": 0,
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Map"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Map not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Map": {
                "title": "Map",
                "description": "This is the record of a Map currently in the map listings of GoFree",
                "properties": {
                    "mapname": {
                        "description": "The name of the map,",
                        "type": "string",
                        "example": "surf_utopia_v3"
                    },
                    "tier": {
                        "description": "The tier of the map,",
                        "type": "integer",
                        "maximum": 8,
                        "minimum": 1,
                        "example": 1
                    },
                    "maxvelocity": {
                        "description": "The limit of speed applied to this map,",
                        "type": "integer",
                        "example": 9999
                    },
                    "ranked": {
                        "description": "Whether this map is currently available in the map listings,",
                        "type": "boolean",
                        "example": true
                    },
                    "record": {
                        "$ref": "#/components/schemas/Time"
                    },
                    "bonuses": {
                        "description": "The amount of bonuses this map has,",
                        "type": "integer",
                        "example": 5
                    },
                    "stages": {
                        "description": "The amount of stages this map has,",
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
            "Time": {
                "title": "Time",
                "description": "This is a time for a map, in a specific style.",
                "properties": {
                    "steamid": {
                        "description": "The SteamID of the player who holds this time,",
                        "type": "string",
                        "example": "STEAM_1:0:12345678"
                    },
                    "mapname": {
                        "description": "The name of the map that this time was set on,",
                        "type": "string",
                        "example": "surf_utopia_v3"
                    },
                    "name": {
                        "description": "The name of the player who holds this time,",
                        "type": "string",
                        "example": "Surfer"
                    },
                    "runtimepro": {
                        "description": "The time, in seconds, it took the player to complete the map,",
                        "type": "integer",
                        "example": 32.6855
                    },
                    "velStartXY": {
                        "description": "The starting speed of the player, using XY velocity, on this map,",
                        "type": "integer",
                        "example": 400
                    },
                    "velStartXYZ": {
                        "schema": "520",
                        "description": "The starting speed of the player, using XYZ velocity, on this map,",
                        "type": "integer"
                    },
                    "velStartZ": {
                        "description": "The starting speed of the player, using only Z velocity, on this map,",
                        "type": "integer",
                        "example": -330
                    },
                    "velEndXY": {
                        "description": "The speed of the player, using XY velocity, at the end of this map,",
                        "type": "integer",
                        "example": 2040
                    },
                    "velEndXYZ": {
                        "schema": "2057",
                        "description": "The speed of the player, using XYZ velocity, at the end of this map,",
                        "type": "integer"
                    },
                    "velEndZ": {
                        "description": "The speed of the player, using only Z velocity, at the end of this map,",
                        "type": "integer",
                        "example": 250
                    },
                    "style": {
                        "description": "The style the player was surfing in when they set this time,",
                        "type": "integer",
                        "example": 250
                    },
                    "time": {
                        "description": "The date the player set this time,",
                        "type": "string",
                        "example": "2021-04-22 17:35:36"
                    }
                },
                "type": "object"
            }
        }
    }
}